requirements.txt
python-dotenv==1.0.1
playwright==1.47.0
pandas==2.2.2
openpyxl==3.1.5
python-dateutil==2.9.0.post0
PyYAML==6.0.2
